+++ stable_baselines3/common/monitor.py
@@ -248,6 +248,8 @@ def load_results(path: str) -> pandas.DataFrame:
             data_frame["t"] += header["t_start"]
         data_frames.append(data_frame)
     data_frames = [df for df in data_frames if not df.empty]
+    if not data_frames:
+        return pandas.DataFrame(columns=["r", "l", "t"])
     data_frame = pandas.concat(data_frames)
     data_frame.sort_values("t", inplace=True)
     data_frame.reset_index(inplace=True)